Improve WIP HPP server
